Fix pcap parsing: check FCS
When reading pcap and passing it on (as 802.15.4 packets) the FCS of the
incoming pcap was not checked. In addition we need to check the length
of incoming packets before stripping the mac header to prevent
inconsistencies.
Tested with a pcap file resulting from capturing with Jackdaw for the
Raven USB stick. These pcap files have a normal checksum and also log
packets with invalid checksum. Relying on the packet contents when
stripping the mac header for packets with invalid checksumm will produce
invalid pcap files on output (resulting in lots of bogus IPv6 addresses
in the foren6 display).
Note that this patch may break processing for other pcap formats (e.g.
resulting from sensniff) which don't store the FCS in the frame but
additional info like RSSI and reserve only a single bit for the FCS
(just valid/invalid).  For this format the changed code won't forward
any packets because they all seem to have wrong FCS. There is currently
no way to automatically distinguish both PCAP formats: Wireshark has a
config option for this. The code needed for the other FCS format is
commented out. Maybe we can come up with an option for selecting the
correct FCS format (on defining the source in foren6?).